refactor: Extracts type definitions from ExtensionsRegistry
